+++ src/components/Buttons/Button.jsx
@@ -1,4 +1,3 @@
-// eslint-disable-next-line react/prop-types
 import { InlineLoadingCircle } from '../index.js';
 
 const Button = ({
