+++ src/components/Buttons/ButtonSecondary.jsx
@@ -5,7 +5,6 @@ const ButtonSecondary = ({
     data = 'click',
     type = 'button',
     bg = 'bg-white',
-    bgHover = 'bg-primary',
     textColor = 'text-black',
     border = 'border border-primary',
     isLoading = false,
@@ -15,7 +14,7 @@ const ButtonSecondary = ({
     return (
         <button
             type={type}
-            className={`${bg} ${textColor} ${border} hover:${bgHover} hover:text-white w-auto rounded px-3 py-2 transition-all ease-in-out hover:cursor-pointer ${className}`}
+            className={`${bg} ${textColor} ${border} hover:bg-[#E7A503] hover:text-white w-auto rounded px-3 py-2 transition-all ease-in-out hover:cursor-pointer ${className}`}
             {...props}>
             {isLoading ? <InlineLoadingCircle /> : data}
         </button>
