+++ src/layouts/Login/pages/AdminLogin.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const AdminLogin = () => {
+    return (
+        <div>AdminLogin</div>
+    );
+};
+
+export default AdminLogin;
