+++ src/layouts/Login/pages/index.js
@@ -1,0 +1,4 @@
+import UserLogin from './UserLogin';
+import AdminLogin from './AdminLogin';
+
+export { UserLogin, AdminLogin };
